organize params and helpers specs into functional specs
